# Refactor `useForcedLayout` to subscribe to store changes and to batch block insertions