--- a/assets/js/blocks/cart-checkout-shared/use-forced-layout.ts
+++ b/assets/js/blocks/cart-checkout-shared/use-forced-layout.ts
@@ -1,20 +1,17 @@
 /**
  * External dependencies
  */
-import {
-	useLayoutEffect,
-	useRef,
-	useCallback,
-	useMemo,
-} from '@wordpress/element';
-import { useSelect, useDispatch } from '@wordpress/data';
+import { useRef, useEffect } from '@wordpress/element';
+import { useRegistry, dispatch } from '@wordpress/data';
 import {
 	createBlock,
 	getBlockType,
 	createBlocksFromInnerBlocksTemplate,
+	BlockInstance,
 } from '@wordpress/blocks';
 import type { Block, TemplateArray } from '@wordpress/blocks';
 import { isEqual } from 'lodash';
+import { MutableRefObject } from 'react';
 
 interface LockableBlock extends Block {
 	attributes: {
@@ -113,114 +110,74 @@ export const useForcedLayout = ( {
 	registeredBlocks: Array< string >;
 	// The default template for the inner blocks in this layout.
 	defaultTemplate: TemplateArray;
-} ): void => {
+} ) => {
 	const currentRegisteredBlocks = useRef( registeredBlocks );
 	const currentDefaultTemplate = useRef( defaultTemplate );
 
-	const { insertBlock, replaceInnerBlocks } =
-		useDispatch( 'core/block-editor' );
+	const registry = useRegistry();
+	useEffect( () => {
+		const { replaceInnerBlocks } = dispatch( 'core/block-editor' );
+		return registry.subscribe( () => {
+			const innerBlocks = registry
+				.select( 'core/block-editor' )
+				.getBlocks( clientId );
 
-	const { innerBlocks, registeredBlockTypes } = useSelect(
-		( select ) => {
-			return {
-				innerBlocks:
-					select( 'core/block-editor' ).getBlocks( clientId ),
-				registeredBlockTypes: currentRegisteredBlocks.current.map(
-					( blockName ) => getBlockType( blockName )
-				),
-			};
-		},
-		[ clientId, currentRegisteredBlocks.current ]
-	);
-
-	const appendBlock = useCallback(
-		( block, position ) => {
-			const newBlock = createBlock( block.name );
-			insertBlock( newBlock, position, clientId, false );
-		},
-		// We need to skip insertBlock here due to a cache issue in wordpress.com that causes an inifinite loop, see https://github.com/Automattic/wp-calypso/issues/66092 for an expanded doc.
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-		[ clientId ]
-	);
-
-	const lockedBlockTypes = useMemo(
-		() =>
-			registeredBlockTypes.filter(
-				( block: Block | undefined ) => block && isBlockLocked( block )
-			),
-		[ registeredBlockTypes ]
-	) as Block[];
-
-	/**
-	 * If the current inner blocks differ from the registered blocks, push the differences.
-	 */
-	useLayoutEffect( () => {
-		if ( ! clientId ) {
-			return;
-		}
-
-		// If there are NO inner blocks, sync with the given template.
-		if (
-			innerBlocks.length === 0 &&
-			currentDefaultTemplate.current.length > 0
-		) {
-			const nextBlocks = createBlocksFromInnerBlocksTemplate(
-				currentDefaultTemplate.current
-			);
-			if ( ! isEqual( nextBlocks, innerBlocks ) ) {
-				replaceInnerBlocks( clientId, nextBlocks );
-				return;
-			}
-		}
-
-		// Find registered locked blocks missing from Inner Blocks and append them.
-		lockedBlockTypes.forEach( ( block ) => {
-			// If the locked block type is already in the layout, we can skip this one.
+			// If there are NO inner blocks, sync with the given template.
 			if (
-				innerBlocks.find(
-					( { name }: { name: string } ) => name === block.name
-				)
+				innerBlocks.length === 0 &&
+				currentDefaultTemplate.current.length > 0
 			) {
+				const nextBlocks = createBlocksFromInnerBlocksTemplate(
+					currentDefaultTemplate.current
+				);
+				if ( ! isEqual( nextBlocks, innerBlocks ) ) {
+					replaceInnerBlocks( clientId, nextBlocks );
+					return;
+				}
+			}
+
+			const registeredBlockTypes = currentRegisteredBlocks.current.map(
+				( blockName: string ) => getBlockType( blockName )
+			);
+
+			const missingBlocks = getMissingBlocks(
+				innerBlocks,
+				registeredBlockTypes
+			);
+
+			if ( missingBlocks.length === 0 ) {
 				return;
 			}
 
-			// Is the forced block part of the default template, find it's original position.
-			const defaultTemplatePosition =
-				currentDefaultTemplate.current.findIndex(
-					( [ blockName ] ) => blockName === block.name
-				);
+			// Initially set as -1, so we can skip checking the position multiple times. Later on in the map callback,
+			// we check where the forced blocks should be inserted. This gets set to >= 0 if we find a missing block,
+			// so we know we can skip calculating it.
+			let insertAtPosition = -1;
+			const blockConfig = missingBlocks.map( ( block ) => {
+				const defaultTemplatePosition =
+					currentDefaultTemplate.current.findIndex(
+						( [ blockName ] ) => blockName === block.name
+					);
+				const createdBlock = createBlock( block.name );
 
-			switch ( defaultTemplatePosition ) {
-				case -1:
-					// The block is not part of the default template so we append it to the current layout.
-					appendBlock( block, innerBlocks.length );
-					break;
-				case 0:
-					// The block was the first block in the default layout, so prepend it to the current layout.
-					appendBlock( block, 0 );
-					break;
-				default:
-					// The new layout may have extra blocks compared to the default template, so rather than insert
-					// at the default position, we should append it after another default block.
-					const adjacentBlock =
-						currentDefaultTemplate.current[
-							defaultTemplatePosition - 1
-						];
-					const position = innerBlocks.findIndex(
-						( { name: blockName } ) =>
-							blockName === adjacentBlock[ 0 ]
-					);
-					appendBlock(
-						block,
-						position === -1 ? defaultTemplatePosition : position + 1
-					);
-					break;
-			}
+				// As mentioned above, if this is not -1, this is the first time we're calculating the position, if it's
+				// already been calculated we can skip doing so.
+				if ( insertAtPosition === -1 ) {
+					insertAtPosition = findBlockPosition( {
+						defaultTemplatePosition,
+						innerBlocks,
+						currentDefaultTemplate,
+					} );
+				}
+
+				return createdBlock;
+			} );
+
+			registry.batch( () => {
+				registry
+					.dispatch( 'core/block-editor' )
+					.insertBlocks( blockConfig, insertAtPosition, clientId );
+			} );
 		} );
-		/*
-		We need to skip replaceInnerBlocks here due to a cache issue in wordpress.com that causes an inifinite loop, see https://github.com/Automattic/wp-calypso/issues/66092 for an expanded doc.
-		 @todo Add replaceInnerBlocks and insertBlock after fixing https://github.com/Automattic/wp-calypso/issues/66092
-		*/
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [ clientId, innerBlocks, lockedBlockTypes, appendBlock ] );
+	}, [ clientId, registry ] );
 };