<!-- Start by describing the changes made in this Pull Request, and the reason for such changes. -->
Following on from the discussion in pca54o-4GY-p2#comment-4658 I implemented the suggestion of @jsnajdr. I altered `useForcedLayout` to subscribe to changes to `core/block-editor` in the registry. I also batched the action dispatches to that store. Whereas we previously inserted blocks individually, we now insert multiple blocks at once if necessary.

I added two new functions (the functionality already existed in the old `useForcedLayout` hook, I just separated it): `getMissingBlocks` and `findBlockPosition`. These functions find which registered blocks are missing from a forced layout, and find the index that these blocks should be inserted at, respectively.

<!-- Reference any related issues or PRs here -->

Fixes #7367


### Testing

#### Automated Tests
* [ ] Changes in this PR are covered by Automated Tests.
  * [ ] Unit tests
  * [ ] E2E tests
#### Internal dev testing

1.  Add WooCommerce Gift Cards, EU VAT Number, and a new plugin generated by the `@woocommerce/extend-cart-checkout-block` template. (Go to `plugins` and exectue: `npx @wordpress/create-block -t @woocommerce/extend-cart-checkout-block my-test-plugin`).
2. Activate them all, and ensure taxes are enabled in WooCommerce (go to WooCommerce -> Settings -> General -> `Enable tax rates and calculations`)
3. Create a new page, add the Cart block. Ensure it loads correctly. Save the page and reload, ensure it's still there.
4. Load this new page on the front-end. Check it renders and works correctly.
5. Create a new page and add the Checkout block. Ensure it loads correctly. Ensure you see the `I want to receive updates about products and promotions.` checkbox in the Contact Information block, ensure you see the VAT Number block rendered at the bottom of the Checkout block (this is intentional).
6. Check the Gift Card boxes appear in the totals sidebar (should see two of them, one checkbox and one "Have a gift card?" panel).
8. Open on the front end and check the VAT Number box is there and the newsletter checkbox. You can also buy and redeem a gift card on your account and check that too, but it's not necessary as long as the "Have a gift card?" panel appears.
9. Repeat these tests on wpcom and ensure nothing breaks!

#### User Facing Testing

<!-- Write these steps from the perspective of a "user" (merchant) familiar with WooCommerce. No need to spell out the steps for common setup scenarios (eg. "Create a product"), but do be specific about the thing being tested. Include screenshots demonstrating expectations where that will be helpful. -->

1. Add the Cart and Checkout blocks to two new pages.
2. Ensure they load correctly on the editor and on the front end.

* [ ] Do not include in the Testing Notes <!-- Check this box if this PR can't be tested by users (ie: it doesn't include user-facing changes or it can't be tested without manually modifying the code). -->

### WooCommerce Visibility

<!-- Check this [this doc](../docs/blocks/feature-flags-and-experimental-interfaces.md) to see if the change is visible in WC core or not (part of the feature plugin or experimental)-->

* [x] WooCommerce Core
* [ ] Feature plugin
* [ ] Experimental

### Performance Impact

<!-- Please document any known performance impact (positive or negative) here. If negative, provide some rationale for why this is an okay tradeoff or how this will be addressed. -->
I tested the performance of my version by logging the difference in `performance.now()` between the start and end of the `subscribe` function (this is where most of the work gets done) and in the old version by logging the difference in `performance.now()` in the `useLayoutEffect`. I took an average of the values of every timestamp collected when inserting a new block and it averaged out around 10ms for both implementations.

I also took the following measurements:

| Metric | Trunk | This branch |
| --- | --- | --- |
| Number of hook calls | 156 | 58 |
| Subscription calls |  (added a subscribe to the registry just to check): 3390 | 434 |
| Heap memory use | Avg: 1026mb | Avg: 479mb |
| Actions dispatched to core/block-editor store | 76 | 67 |

### Changelog

> Skipping