@senadir 2022-12-14
I love the effort you put into testing the performance of this Thomas!