@opr 2022-12-14
> I successfully tested this PR both using [wp.org](https://wp.org/) and [wp.com](https://wp.com/).
> 
> I do have to say that on [wp.com](https://wp.com/), I haven't been able to run `npx @wordpress/create-block -t @woocommerce/extend-cart-checkout-block my-test-plugin`. When trying to do that via SSH, I received the following error message:
> 
> ```shell
> bash: npx: command not found
> ```
> 
> Apart from that, the PR was working as expected.

No worries Niels, you can run that locally, cd into the new directory, run `npm run build` then upload the zip that gets created.

Thank you for testing @nielslange and @senadir