@ralucaStan 2022-12-30
Great work on this @opr, hopefully we won't have any more reports of https://github.com/woocommerce/woocommerce-blocks/issues/5866